+++ lib/features/profile/version_control_screen.dart
@@ -84,6 +84,7 @@ class VersionControlScreen extends StatelessWidget {
 
   List<String> _notesFor(AppRelease release, AppLocalizations l) {
     return switch (release.label) {
+      '1.0.8+9' => [l.release108Change1, l.release108Change2],
       '1.0.7+8' => [
         l.release107Change1,
         l.release107Change2,